fix: adjust column types in migration

+++ src/main/resources/db/migration/V3__create_invoice_item.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS invoice_item (
     id BIGSERIAL PRIMARY KEY,
     description VARCHAR(255) NOT NULL,
-    quantity NUMERIC(12,2) NOT NULL,
+    quantity BIGINT NOT NULL,
     unit_price NUMERIC(12,2) NOT NULL,
     vat_rate NUMERIC(5,2) NOT NULL DEFAULT 23.00,
     net_value NUMERIC(12,2) NOT NULL,
